Switch to formatter-maven-plugin for code formatting (closes #125)
git-code-format-maven-plugin is based on google-java-format java format for code formatting. But when running on JDK 16 and newer, due to JEP 396: Strongly Encapsulate JDK Internals by Default, it requires adding multiple --add-exports to be set. This is an issue because --add-exports is not a valid option on JDK 8 and using -XX:+IgnoreUnrecognizedVMOptions does not help.

We need to continue running our tests against JDK 8, so instead of git-code-format-maven-plugin we will use formatter-maven-plugin 2.16 (2.17+ requires java 11+) for code formatting.

Code have been reformatted in order to adhere to Eclipse code formatting rules.